refactor(frontend): extract topbar action types into navigation/topbarTypes
Move TopbarContextAction and RootLayoutOutletContext out of App.tsx into
src/navigation/topbarTypes.ts so hooks and pages no longer import types
from the root component. App.tsx re-exports both types for
compatibility; all internal importers repointed.
